+++ src/shipbattle.h
@@ -1,15 +1,15 @@
 #pragma once
 #include "types.h"
   
-extern const int PLAYER_Y_LEVEL;
-extern const int ENEMY_Y_LEVEL;
-extern const int ENEMY_FIRE_SPEED;
-extern const int PLAYER_FIRE_SPEED;
-extern const int PLAYER_MOVE_SPEED;
-extern const int ENEMY_MOVE_SPEED;
-extern const int SHOT_MOVE_SPEED;
-extern const int PLAYER_SHIP_SIZE;
-extern const int ENEMY_SHIP_SIZE;
+extern const int8_t PLAYER_Y_LEVEL;
+extern const int8_t ENEMY_Y_LEVEL;
+extern const int8_t ENEMY_FIRE_SPEED;
+extern const int8_t PLAYER_FIRE_SPEED;
+extern const int8_t PLAYER_MOVE_SPEED;
+extern const int8_t ENEMY_MOVE_SPEED;
+extern const int8_t SHOT_MOVE_SPEED;
+extern const int8_t PLAYER_SHIP_SIZE;
+extern const int8_t ENEMY_SHIP_SIZE;
 
 void initializebattle(ShipBattleData* shipbattledata);
 void updatebattle(ShipBattleData* shipbattledata, GameData* gamedata);
